Pull out folder scenario test from segment scenarios Use timebased caseref instead of random number
