+++ src/functionalTest/java/uk/gov/hmcts/reform/em/hrs/FolderScenarios.java
@@ -1,0 +1,22 @@
+package uk.gov.hmcts.reform.em.hrs;
+
+
+import org.junit.Test;
+
+import static org.hamcrest.Matchers.empty;
+import static org.hamcrest.Matchers.equalTo;
+
+public class FolderScenarios extends BaseTest {
+
+
+    @Test
+    public void shouldCreateFolderWhenDoesNotExistAndReturnEmptyFileNames() {
+        final String nonExistentFolder = "audiostream000000";
+
+        getFilenames(nonExistentFolder)
+            .assertThat().log().all()
+            .statusCode(200)
+            .body("folder-name", equalTo(nonExistentFolder))
+            .body("filenames", empty());
+    }
+}
