+++ src/functionalTest/java/uk/gov/hmcts/reform/em/hrs/ShareHearingRecordingScenarios.java
@@ -25,7 +25,7 @@ public class ShareHearingRecordingScenarios extends BaseTest {
     @Before
     public void setup() throws Exception {
         createFolderIfDoesNotExistInHrsDB(FOLDER);
-        caseRef = randomCaseRef();
+        caseRef = timebasedCaseRef();
         filename = filename(caseRef);
 
         int cvpBlobCount = testUtil.getCvpBlobCount(FOLDER);
